+++ foodplanner/lib/pages/settings/settings.dart
@@ -10,7 +10,7 @@ import 'package:flutter_advanced_switch/flutter_advanced_switch.dart';
 import 'package:flutter_advanced_segment/flutter_advanced_segment.dart';
 import 'package:foodplanner/pages/settings/administrate_children.dart';
 import 'package:foodplanner/pages/settings/admin_approve_page.dart';
-import 'package:foodplanner/pages/settings/SchoolClasses.dart';
+import 'package:foodplanner/pages/settings/school_classes.dart';
 import 'package:foodplanner/routes/user_roles.dart';
 import 'package:provider/provider.dart';
 import 'package:foodplanner/pages/settings/deactivate_accounts.dart';
